Restructure and complete the functionality of blue/green deployment
